delete nested classes and make separate file for each class, add new diagram
