--- Animal/Animal/ClassDiagram1.cd
@@ -1,126 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<ClassDiagram MajorVersion="1" MinorVersion="1">
-  <Class Name="Animal.Birds">
-    <Position X="0.5" Y="3.5" Width="1.5" />
-    <Compartments>
-      <Compartment Name="Nested Types" Collapsed="false" />
-    </Compartments>
-    <NestedTypes>
-      <Class Name="Animal.Birds.Preybirds">
-        <Compartments>
-          <Compartment Name="Nested Types" Collapsed="false" />
-        </Compartments>
-        <NestedTypes>
-          <Class Name="Animal.Birds.Preybirds.Eagle">
-            <TypeIdentifier>
-              <NewMemberFileName>Classes\Birds.cs</NewMemberFileName>
-            </TypeIdentifier>
-          </Class>
-        </NestedTypes>
-        <TypeIdentifier>
-          <NewMemberFileName>Classes\Birds.cs</NewMemberFileName>
-        </TypeIdentifier>
-      </Class>
-    </NestedTypes>
-    <TypeIdentifier>
-      <HashCode>AAAAAAgAEAAAAAAAAAAAAAAAAAAAAAAAAAABAAAAAAA=</HashCode>
-      <FileName>Classes\Birds.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="Animal.Fish">
-    <Position X="2.5" Y="3.5" Width="1.5" />
-    <Compartments>
-      <Compartment Name="Nested Types" Collapsed="false" />
-    </Compartments>
-    <NestedTypes>
-      <Class Name="Animal.Fish.Whale">
-        <TypeIdentifier>
-          <NewMemberFileName>Classes\Fish.cs</NewMemberFileName>
-        </TypeIdentifier>
-      </Class>
-      <Class Name="Animal.Fish.BlueWhale">
-        <TypeIdentifier>
-          <NewMemberFileName>Classes\Fish.cs</NewMemberFileName>
-        </TypeIdentifier>
-      </Class>
-    </NestedTypes>
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAA=</HashCode>
-      <FileName>Classes\Fish.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="Animal.Insects">
-    <Position X="4.5" Y="3.5" Width="1.5" />
-    <Compartments>
-      <Compartment Name="Nested Types" Collapsed="false" />
-    </Compartments>
-    <NestedTypes>
-      <Class Name="Animal.Insects.FlyingBugs">
-        <Compartments>
-          <Compartment Name="Nested Types" Collapsed="false" />
-        </Compartments>
-        <NestedTypes>
-          <Class Name="Animal.Insects.FlyingBugs.Bee">
-            <TypeIdentifier>
-              <NewMemberFileName>Classes\Insects.cs</NewMemberFileName>
-            </TypeIdentifier>
-          </Class>
-          <Class Name="Animal.Insects.FlyingBugs.Flyingbee">
-            <TypeIdentifier>
-              <NewMemberFileName>Classes\Insects.cs</NewMemberFileName>
-            </TypeIdentifier>
-          </Class>
-        </NestedTypes>
-        <TypeIdentifier>
-          <NewMemberFileName>Classes\Insects.cs</NewMemberFileName>
-        </TypeIdentifier>
-      </Class>
-    </NestedTypes>
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAEAAAAAAAAAAAAAAAAAAAAAAAAAABAAAAAAA=</HashCode>
-      <FileName>Classes\Insects.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="Animal.Mammals">
-    <Position X="6.5" Y="3.5" Width="1.5" />
-    <Compartments>
-      <Compartment Name="Nested Types" Collapsed="false" />
-    </Compartments>
-    <NestedTypes>
-      <Class Name="Animal.Mammals.BigCats">
-        <TypeIdentifier>
-          <NewMemberFileName>Classes\Mammals.cs</NewMemberFileName>
-        </TypeIdentifier>
-      </Class>
-    </NestedTypes>
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAAQAAAAAAAAAAAAAAAAAAAAAAAAACAAAAAAA=</HashCode>
-      <FileName>Classes\Mammals.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="Animal.Reptile">
-    <Position X="8.5" Y="3.5" Width="1.5" />
-    <Compartments>
-      <Compartment Name="Nested Types" Collapsed="false" />
-    </Compartments>
-    <NestedTypes>
-      <Class Name="Animal.Reptile.Lizard">
-        <TypeIdentifier>
-          <NewMemberFileName>Classes\Reptile.cs</NewMemberFileName>
-        </TypeIdentifier>
-      </Class>
-    </NestedTypes>
-    <TypeIdentifier>
-      <HashCode>AAAAAAgAEAAAAAAAAAAAAAAAAAAAAAAABAAAAAAAAAA=</HashCode>
-      <FileName>Classes\Reptile.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Class Name="Animal.Animal">
-    <Position X="4.5" Y="0.5" Width="1.5" />
-    <TypeIdentifier>
-      <HashCode>AAAAAAAAEQAAAAAAAAAAAAQAAACAAAAABAACAAAAAAA=</HashCode>
-      <FileName>Classes\Animals.cs</FileName>
-    </TypeIdentifier>
-  </Class>
-  <Font Name="Segoe UI" Size="9" />
-</ClassDiagram>
