+++ Animal/Animal/Classes/BlueWhale.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Animal.Classes
+{
+    public class BlueWhale : Whale // class child bluewhale inherit from parent whale  
+    {
+        public override string Color()
+        {
+            return "im whale my color is blue";
+        }
+
+    }
+}
