+++ Animal/Animal/Classes/Preybirds.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Animal.Classes
+{
+    public abstract class Preybirds : Birds // abstraction
+    {
+        public abstract void EyeColor();
+
+    }
+}
