+++ Animal/Animal/Classes/Reptile.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace Animal
 {
     // child reptile inherit from parent animal and i legs num interface    
-    public class Reptile : Animal
+    public class Reptile : Animal,ILegsNum
     {
         public new string Sound()
         {
@@ -22,17 +22,17 @@ namespace Animal
         {
             return "I have four legs";
         }
-        // child lizzard imhrit from parent reptile
-        public class Lizard : Reptile
-        {
-            public new string Sound()
-            {
-                return "im a lizard, my sound is hissing,";
-            }
-            public new string Tail()
-            {
-                return "i have tail";
-            }
-        }
+        //// child lizzard imhrit from parent reptile
+        //public class Lizard : Reptile
+        //{
+        //    public new string Sound()
+        //    {
+        //        return "im a lizard, my sound is hissing,";
+        //    }
+        //    public new string Tail()
+        //    {
+        //        return "i have tail";
+        //    }
+        //}
     }
 }
