+++ Animal/Animal/Classes/Whale.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Animal.Classes
+{
+    public abstract class Whale : Fish // abstraction
+    {
+        public abstract string Color();
+    }
+}
